+++ crates/store/src/accounts/mod.rs
@@ -29,6 +29,10 @@ mod tests;
 /// Convenience for an in-memory-only account tree.
 pub type InMemoryAccountTree = AccountTree<LargeSmt<MemoryStorage>>;
 
+#[cfg(feature = "rocksdb")]
+/// Convenience for a persistent account tree.
+pub type PersistentAccountTree = AccountTree<LargeSmt<miden_crypto::merkle::smt::RocksDbStorage>>;
+
 // HISTORICAL ERROR TYPES
 // ================================================================================================